added some reports and tests params

+++ iron-infra/src/main/java/com/ironsource/mobile/MobileSO.java
@@ -134,7 +134,7 @@ public class MobileSO extends SystemObjectImpl {
 	 * 
 	 * @throws Exception in case of timeout or "RS" with "E" was reported.
 	 */
-	public void waitForRSCode(RSCode rsCode, FlowCode flowCode, int timeout) throws Exception {
+	public void waitForRSCode(RSCode rsCode, FlowCode flowCode, long timeout) throws Exception {
 		long now = System.currentTimeMillis();
 		boolean exist = false;
 		report.report("waiting for RS Code '"+ rsCode.getRsCode() +"' for Flow " + flowCode.getFlowCode() + "with timeout of " + timeout + " millis");
